Mask rcnn inference (#6)
* Fixes to get weights from Mask R-CNN

* Fixes when number of boxes is zero

* Add initial working version of Mask R-CNN

Still need to add better post-processing that resizes the images

* Add projection of masks in image using GridSampler

Gives slightly different results than the approach from Detectron. Need to verify if that matters. Also, it seems that Ross' proposal can't handle having different branches computed using different boxes without too much hacking. One thing that I'm thinking now is that I'm following a lua-torch approach, and I'm trying to avoid having the user write too much code in the forward, given the complexity of detection models. This should be double-checked to see if it's actually something we want to do.

* Bugfixes and add GPU kernel for compute_flow

Can optimize the kernel but it is already fairly ok

* Bugfix

* Move check of empty tensor

* Bugfix

* Fix kth value + add extra arguments to PostProcessor

* Remove compile decorator

It's being removed from pytorch, and on real use-cases it seemed to slowdown significantly the runtime, because currently every different tensor size gets compiled on the fly

* Add CUDA implementation of NMS

Inspired from py-faster-rcnn. This drastically improves runtime performance on FPN models, as the NMS was one of the single CPU operation being performed in the pipeline, making the GPUs starve

* Fix dimension computations in compute_flow

* Add initial version of inference file

Just to obtain some runtime and accuracy numbers. This needs to be made more generic

* Add support for Mask R-CNN in Ross proposal

* Bugfix

* Add no abstraction Mask R-CNN model

* Remove comments

* Make padding for FPN more generic

Ideally, we would replace the scale=2 with size, but this results in a drop in mAP for the pre-trained models. I'll investigate if it changes anything once we train the model from scratch

* Add TODO