Apply linter
Summary:
While working on another diff, I noticed that the linter has not been run recently on a lot of files. This diff only contains the changes from running `./infra/linter.sh` on a devgpu.

(I have not confirmed, but it may be the case that the changes made by `./infra/linter.sh` on a devfair machine are different. If so, we should probably identify the exact versions of isort, black, and flake8 that we want to run.)

Reviewed By: ppwwyyxx

Differential Revision: D13695781

fbshipit-source-id: 46bdbfe879528fbeb903d301ea02510da79d280f