Refactor FPN module
Summary: The existing FPN code is hard to read because it uses a lot of unnecessary `getattr` calls and uses names that don't align with the description in the paper (though some are bad names inherited from C2 Detectron). This diff does some cleanup to make the code more readable.

Reviewed By: ppwwyyxx

Differential Revision: D13741706

fbshipit-source-id: a566984ec692355ac1b0b1b55a0b3bc45d432c40