tcp port for multiprocess spawn is determined based on the user (#80)
Summary:
The fixed port for the spawn function prevents you from launching several instances of training on the same machine. This is actually a useful feature since sometimes after a crash not all processes are dead. So, that is how you know that you need to find and kill them.

On the other hand, fixed port prevent several users to launch their training scripts on the same machine simultaneously. I believe that in same cases it is desirable, so the port can be defined based on the user.
Pull Request resolved: https://github.com/fairinternal/detectron2/pull/80

Differential Revision: D13749913

Pulled By: alexander-kirillov

fbshipit-source-id: 5457c3170cdeadbacdb04d8e3677575f86f6e514