Create a backbone interface
Summary: This diff encapsulates the backbone in a consistent way: now it's always an instance of a subclass of the `Backbone` ABC instead of being an `nn.Sequential` container with different key, value pairs depending on the backbone case. This consistent design makes it much easier to access properties like feature map strides and channel counts from other parts of the code (like RPN and ROIHeads).

Reviewed By: ppwwyyxx

Differential Revision: D13747985

fbshipit-source-id: a1d8736a4f4f2b76a4de8bbe1cb0037dbe31b236