Refactor AnchorGenerator API
Summary:
The `AnchorGenerator` API was a little awkward because it had asymmetric treatment of anchor sizes vs. aspect ratios. For sizes, you could specify a set of sizes for a single feature map or only one size for each of multiple feature maps. For aspect ratios, you could specify one or more, but they always had to be the same for each feature map. These restrictions are artificial.

This diff makes the treatment symmetric and expands the available functionality: you can now specify any combination of anchor sizes and aspect ratios for each feature map by giving a list of lists of sizes (or aspect ratios). To make the case in which the same size or aspect ratio is used for all feature maps easier to specify (as in FPN), we use broadcast semantics.

Reviewed By: ppwwyyxx

Differential Revision: D13829594

fbshipit-source-id: 6ad011770601308a9c3df5de06bca4bd7d95bb68