Rename BoxCoder
Summary: A class named `BoxCoder` was carried over from maskrcnn_benchmark. Its role is to compute bounding box regression deltas and to all apply them. The name `BoxCoder` is misleading; its role is not to "encode" and "decode" boxes, which suggests changing the representation/format, but not the geometric properties of a box. Instead, the class represents a transformation from one box to another, where the transformation is parameterized by 4 "deltas" that describe how to shift and scale a box.

Reviewed By: ppwwyyxx

Differential Revision: D13842159

fbshipit-source-id: 8a0c346e687d714791effdfaf9dd2f7a3919af34