Use 1 im/gpu during inference
Summary:
While doing something unrelated, I noticed that inference results vary slightly depending on if I was using 1 or 2 GPUs while testing with `infra/fb/run_inference_tests.sh`. The issue is due to using > 1 image per GPU during inference. When > 1 image is used, some zero padding is usually required. This zero padding creates small differences in the output and final AP.

To avoid these differences, this diff forces inference to always use 1 image per GPU. This is also preferred since the standard fair comparison in papers on timing is with 1 image per GPU (no batching).

While digging, I also found that there are cases in which inference is run more than once on the same image due to how the `DistributedSampler` pads its index sampling. This duplication leads to a json-results-file that has duplicate results entries for a small number of images. I don't think this causes issues with the COCO evaluation, but I've fixed it on our end to be safe.

Reviewed By: wanyenlo, ppwwyyxx

Differential Revision: D13938286

fbshipit-source-id: 8810aa385419a9a0b5ff51b3ea0cb6005f5ab194