Remove "roidb" name
Summary:
This is another diff that's primarily focussed on naming. The target this time is `roidb`.

This term, meaning "region of interest database", dates back to the original matlab R-CNN code in which it was designed primarily to store precomputed object proposals ("rois"), but also held metadata about each image and its annotations in the dataset. (And I think I took the term from some of Andrea Vedaldi's code.)

The term isn't very meaningful anymore given that `roi` isn't the primary feature of each `roidb`. For each dataset image it:
- holds metadata about that image (height, width, path to file)
- holds annotations about the objects in that image
- annotations might be boxes, segmentation masks, keypoints, etc
- in the future it will also store pixel-wise label maps for semantic segmentation

In my renaming I'm using `dataset_dict` to reflect that it's a plain dict that stores many types of information about a dataset image and doesn't have an API beyond key, value access. Suggestions for better names are always welcome.

I've also done a little bit of additional cleanup and tried to make it clear when a `dataset_dict` is used vs. a similar dict structure that's used for storing inference predictions.

Reviewed By: wanyenlo, ppwwyyxx

Differential Revision: D13938178

fbshipit-source-id: eb6b7a6badb7bab6dca717b64541f9af9990ee4f