Add tensorboardX writer.
Summary:
Add a monitor which uses https://github.com/lanpa/tensorboardX to write data to
tensorboard-compatible format.

tensorboardX is not a required dependency, so the import is delayed into the class.

We don't seems to have the mirror of tensorboardX in fbcode now (there is a fork in
pytorch/tensorboardX).

Reviewed By: wanyenlo

Differential Revision: D13977304

fbshipit-source-id: 372dbc35f12b16be7249a24b65eebab41e8ebbb1