Fix speed computation of evaluation
Summary:
The size used in the computation was `len(dataset)` in the past,
and was then changed to `len(data_loader)` in D14096497.

As a result the speed number becomes 8x larger if running on 8 GPUs.

Given that different workers may have different number of images for inference,
it's more accurate to compute the speed based on only one worker.

Reviewed By: rbgirshick

Differential Revision: D14164168

fbshipit-source-id: e620f62a9f696bede299150024fe149baf785366