Register box & mask heads.
Summary:
The keypoint head used to parse the input size from `BACKBONE.OUT_FEATURE_CHANNELS`.

However in our existing models, the same strategy does not apply to mask head: in C4 models, the number of input channels to mask head is the output channels of the Res5 head.
This diff makes mask head use `COMPUTED_INPUT_SIZE`, and for consistency I make the box head & keypoint head do the same (though not necessary, and I also hope to introduce as few computed attributes as possible).

In general I think this is an issue with "reading configs of a different component". When A reads a config value from `cfg.B.XX`, it is assuming that `A` will be used together with `B` in certain way, and this assumption may not hold. Not sure if there are more general solutions for this.

Reviewed By: wanyenlo, rbgirshick

Differential Revision: D14351360

fbshipit-source-id: 89a742a280a5d4c2c453dea33b76cc21b8ae3972