Use per-dataset metadata; Associate metadata with dataset splits.
Summary:
The existing code uses a globally unique "COCOMeta":
this will make it hard to use >1 datasets together.
It also makes it hard to hard-code some metadata (as requested in https://github.com/fairinternal/detectron2/issues/35)

In this diff, the data loading code uses per-dataset metadata:
each "Metadata" is associated with a name (e.g., "coco") and different names
can then map to different "Metadata".

Also, inside DatasetCatalog, I make each "dataset split" (e.g., "coco_2014_train")
also map to a "dataset name" (e.g., "coco").
This change will make the data loader more human-friendly, for example, now we
will be able to print per-category statistics or visualization of certain "dataset split".

Note that to do this we stop calling "coco_2014_train" the "name of a dataset",
but the "name of the split" instead. The accurate meaning is a name that corresponds to a list of dicts. The naming might be confusing. Maybe `register(key, dataset_name, func)` instead?

An alternative way doing this: let each "split" correspond to its own metadata or a shared metadata, without the need for `MetadataCatalog`?

Note that the metadata does not always come automatically from data loading, but could come from other code (or hard-coded like COCO). So in the general case the two approaches are equally complicated for a user: user will still need to provide the mapping from name to metadata.
Having a catalog for metadata makes it possible to access metadata without the need to touch an actual "split" which is not always available: the metadata will be indexed by "coco", rather than "coco_2014_train". And I suppose this will be useful for demos, predictions, visualizations, when the actual dataset is not on disk.

Reviewed By: rbgirshick

Differential Revision: D14386541

fbshipit-source-id: fe2532abdd2be4cb2413191ddd587fc0a7005fbd