move fast_rcnn_inference inside FastRCNNOutputs to allow easier patch
Summary:
I'm writing the machanism for patching caffe2 equivalent `fast_rcnn_inference`, and find it's a bit hard to do because the fast_rcnn_inference is coupled with FastRCNNOutputs. (the root reason is actually because we want to avoid splitting tensors in c2 path since it might be tricky for tracing).

A solution is moving the `fast_rcnn_inference` call from within ROIHeads to `inference` member function of FastRCNNOutputs, after this change only `inference` needs to be taken care of. This change also hides `fast_rcnn_inference` from RoIHeads and makes FastRCNNOutputs a bit more self contained.

Reviewed By: rbgirshick

Differential Revision: D14582249

fbshipit-source-id: 9efe517429fbdab63e5077dedb37ae423388bdc3