Remove batch collator.
Summary:
The reason we use "batch" in deep learning is for efficiency.
So when it does not brings efficiency gains, but only more complexity, it's time to avoid it.

The existing batch collator does not really batch inputs into tensors, but a list, due to the nature
of detection tasks where inputs do not have regular shapes. As a result there isn't really a performance gain.
The batching of images is necessary, but it's equally efficient to do it inside the model (v.s.
inside the data loader, see last diff D14623370).

Before, an user working on new types of data will need to provide new dicts, new `Transform` and a new `BatchCollator`.
Now this extra level of complexity (`BatchCollator`) is turned into a trivial identity function.
As a result it will be easier to support new data (easier in terms of interface. In terms of code users still need to write code in the `Transform`).

After this I'll make both the `Transform` and `BatchCollator` arguments to
`build_detection_xxx_loader`, but `BatchCollator` defaults to the trivial one.
If some types of data is performance-sensitive and still needs to be batched inside dataloader,
users can still provide a custom `BatchCollator`.

Reviewed By: rbgirshick

Differential Revision: D14627686

fbshipit-source-id: cb77a9e41b61130ccad4ab5210dcc52d759c8c1f