Use "Evaluator" for evaluation. (#129)
Summary:
Like hooks/callbacks, but for the evaluation loop.
With this, we'll be able to add evaluation for cityscapes and other datasets easier.

We want to avoid a generic evaluator function which takes all the results, because the results may not fit into memory (it already happens in COCO which is why we encode them into RLE). Instead we use a class with a method to process each single result.

We explicitly avoid generic training loops with hooks/callbacks, but it sounds like a good idea for evaluation loop.
Because the concept of evaluation loop is much simpler: in the current design, it only has the concept of "input", "output", and a function (predictor) that maps input to output. That's all the assumptions made here.

To support test-time augmentation, we just need to wrap the trained network into a predictor which, augments the input and ensembles all the outputs.

Pull Request resolved: https://github.com/fairinternal/detectron2/pull/129

Reviewed By: rbgirshick

Differential Revision: D14570618

Pulled By: ppwwyyxx

fbshipit-source-id: 2eb2fbfdcb2898714b15d15512ae4baf5556a741