Add per-split Metadata
Summary:
We used to store COCO-specific information in DatasetCatalog, which is hacky and do not work for more formats.
It is implemented back then to avoid accessing `dataset.ds.coco_api` in evaluation, which was more (or equally) hacky.
Now this is hopefully a not-hacky solution.

In addition to the per-dataset metadata, now each split can have its own metadata, and can include basic
primitives such as the path to the split.
Then this information will be easily accessible in evaluation or visualization.

After this change, the `DatasetCatalog` class becomes indepdent of detection -- it's general enough to
work in semantic segmentation / classification, etc. I wonder whether to move it outside `detection/` later.
I think it's a good plan to gradually refactor things so that they are more general and move them to shared code.

Reviewed By: wanyenlo, rbgirshick

Differential Revision: D14656431

fbshipit-source-id: c55320278e84a84d4e6e104d7439f1ca5ee264ab