Cache files locally
Summary: As Yuxin Wu pointed out, the previous model caching diff only handled models from the `ModelCatalog`. This diff extends it to handle all models. Cache-on-model-load is enabled by the user with a new config option. It is off by default as this behavior is only likely useful in cases of slow network attached storage.

Reviewed By: ppwwyyxx

Differential Revision: D14836399

fbshipit-source-id: 3a08353b362c7c734bcaf7ebe270c3954341e0bf