Keep all anchors by default
Summary: The previous default, based on the original Faster R-CNN implementation, removed all anchors that have any area outside of the image. This seems to not be useful has caused problems at least once when trying to implement other papers. This diff retains the functionality but makes the default to keep all anchors.

Reviewed By: ppwwyyxx

Differential Revision: D14854961

fbshipit-source-id: 569eec4d332e1496d9938ab89e1cd3e28111edad