fsync results file and avoid divide by zero
Summary:
This diff addresses two issues:

1) If a job dies after training finishes and then is retried, it leads to a divide by zero in `do_train()`
2) Attempt to address the issue that surfaced (1)

Chronology:

I ran baseline run f110017597, which crashed in COCO evaluation. Subsequent retries died with a divide-by-zero exception. The original crash was due to a COCO results json parsing error.

After fixing the divide-by-zero, I tried re-running the job as new job f110171365. This new job crashed with the same json parsing error as the first job. To investigate the issue, I tried to load the json results file in a jupyter console. This succeeded, indicating that the file contents are ok and the problem seems to be related to flushing the file content to disk. Using `os.fsync` is an attempt to make this process more robust on the cluster (https://docs.python.org/3/library/os.html#os.fsync). Rerunning the original job as new job f110185121 succeeded with the fsync change (though it could also be that the underlying issue was transient and went away; I have never seen it before).

Reviewed By: sampepose

Differential Revision: D15028477

fbshipit-source-id: c6fe89949751cfc56c222f170158521e7c2a4be4