Make L1 loss default for all box regression
Summary: Xinlei observed in some TensorMask comparisons that using L1 loss consistently outperforms Smooth L1 in terms of COCO AP. In the breakdown of AP by IoU threshold, one can see that AP@50 is systematically better with Smooth L1, but at higher IoU thresholds t, plain L1 gives better AP@t. The choice of Smooth L1 was based on AP@50 in the time of PASCAL AP (so it's good to see that it is actually better in that regime). However, the overall average (used by COCO) is in favor of plain L1 loss. Since COCO is dominant, and I think we should prioritize AP at higher IoU thresholds regardless of COCO, this diff changes the default from Smooth L1 to plain L1.

Reviewed By: wanyenlo, ppwwyyxx

Differential Revision: D14929287

fbshipit-source-id: 00d94ca59ad1feaa8ca8606860c5322101a6bf02