Cleanup anchors
Summary:
Regarding BC: saved models actually store the anchors they were trained with and when loaded will use those anchors regardless of what code is implemented. Due to that, changing the logic for anchor computation does not break BC.

Given this, should we bother to keep the old code for the original RPN anchor definitions? I think I'm slightly in favor of just deleting it and only using the anchor definition going forward.

Reviewed By: ppwwyyxx

Differential Revision: D14945174

fbshipit-source-id: 904bd10b39a73e859b3835e81ce344b675bd01f2