fix false negative calculation
Summary: It looks like the logged false negative rate per minibatch is incorrect since we moved the bg class index from `0` to `num_fg_classes`. This diff fixes the calculation and also renames `num_fg_classes` to `bg_class_ind`, which I think is semantically more meaningful in the context in which it's used.

Reviewed By: ppwwyyxx

Differential Revision: D15804919

fbshipit-source-id: 6c1a916492032f742b4a39764fd33d0a6218ae02