move mask/keypoints logic to separate methods in StandardROIHeads
Summary:
This makes it easier to subclass `StandardROIHeads`.

Now they both take all sampled proposals as input.
When both mask/keypoint is on, the extra call to `select_foreground_proposals` is unnecessary,
but I'd like to consider it as a light cost of abstraction..

Controversial next steps:
`init_{mask,keypoint,box}_head(self, cfg)` as well?

Reviewed By: rbgirshick

Differential Revision: D16134395

fbshipit-source-id: a4fee4a27414ba4b7d3b6bab524999048ff9ff65