Limit GPU memory usage in mask pasting
Summary: I've been doing some experiments with high-res images (e.g., 2000x4000) and in this case we run out of GPU memory when pasting masks. This batches the pasting step into chunks so that no more than 1GB of GPU memory is used for pasting. The 1GB choice is kind of arbitrary (it's HUGE for the standard COCO setting).

Reviewed By: ppwwyyxx

Differential Revision: D16161060

fbshipit-source-id: dd8e95d06af0548cf9d75086ca71813216d3e8ab