Fix out of bounds issue in coco_evaluation
Summary: Previous implementation assumed the eval_params to have at least three max_dets and it directly tried to access the third element of that corresponding array. In case number of classes was configured to 1 it would accidentally work since it returned early a few lines above. This change instead reads the last max_det which should always work.

Reviewed By: ppwwyyxx

Differential Revision: D16365181

fbshipit-source-id: f9971ac82e19732d736d9259e952a39a697ed61b