rewrite the image transformation interface
Summary:
Note: This diff does not make "DetectionTransform" easier. It aims to resolve design issues of the
underlying image transformation code.

Name suggestions and comments needed.

"TransformOp" performs deterministic operation that runs on a set of pre-defined primitive data types
(boxes, images, segms, coords).
This should be quite general.
(New data type always needs to implement its own transformation if it cannot be achieved with the primitives.
Don't think there is a straightforward solution to that.)

"ImageTransformer" represents the (likely random) policy that generates such ops.
It implicitly makes an assumption that it only needs to look at the image to generate the transform.

More about this assumption:
This assumption exists because anything more complicated than it (e.g. crop) will start to depend on the actual task and
what's available in the dataset.

Crop, as an example, needs to look at both images and instances to decide where to actually crop.
This can be implemented on the user's side by creating an "CropOp" with custom rules together with
the ops created by the "ImageTransformers".
Or we may as well remove the concept of "ImageTransformers" completely and implement all policies individually.
But the benefit of keeping this concept is to be able to configure them in a more descriptive list, e.g.:
```
INFO [07/12 18:21:25] data.detection_transforms: Transformers used in training: [ResizeShortestEdge(short_edge_length=(800,), max_size=1333, sample_style='choice'), RandomFlip()]
```
We can see which one is better after including Alex's diff.

Reviewed By: wanyenlo, rbgirshick

Differential Revision: D16234586

fbshipit-source-id: aaafecb98c51068cc15c3a653caaa3c028e7f8b0