matcher reform
Summary: Reform to matcher. Instead of defining high_threshold and low_threshold, the class accepts as inputs a list of thresholds and a list of labels corresponding to the threshold ranges. The labels are then used to define positive, background and ignore regions/predictions in rpn, roi_heads and retinanet.

Reviewed By: gkioxari

Differential Revision: D16206923

fbshipit-source-id: 1bd65e52c3a20ea7033bde9b384c35ba25ccb56a