Refactor PathManager for I/O
Summary:
This adds more I/O functionality to `PathManager`:
- exists / isfile
- ls
- open

I also renamed `_get_file_name` -> `_get_local_path` to make it more clear that the return path is guaranteed to be accessible with native OS calls. In the case of a remote URI, this results in downloading and caching the resource.

Reviewed By: ppwwyyxx

Differential Revision: D16545531

fbshipit-source-id: 5f085108fa68bcc8f9e2e65bcb5387aee9686b96