Resume training at the correct iteration
Summary:
As noted by Xinlei Chen in https://github.com/fairinternal/detectron2/issues/214, if training is "resumed" after it has finished (this can happen if there was a failure in evaluation and the job retries) then one additional iteration is run.

Actually, whenever we resume from a checkpoint we run an extra training iteration. The "iteration" field stored in the checkpoint holds the value of the last completed iteration and therefore when resuming from a checkpoint the start iteration should be the checkpoint iteration + 1.

Reviewed By: ppwwyyxx

Differential Revision: D16756133

fbshipit-source-id: 374e4a1f57bcee80e4ef369a1b3f70bc144e21da