Normalize keypoint scores by pooled size. (#252)
Summary:
The keypoint probabilities are produced by a softmax over the HxW region, where H, W are the absolute size of the box.
As a result, __keypoints in large boxes tend to have lower probabilities__.

This does not affect mAP computation, since the scores are unused there.
But it affect interpretation of the results. When running a webcam demo, I found large objects have lower scores and are filtered out by the threshold.

This diff normalize it using the sum of scores in the pooled feature map instead. As a result the scores do not sum to 1 any more. Not sure whether this makes sense.
Pull Request resolved: https://github.com/fairinternal/detectron2/pull/252

Reviewed By: jonmorton

Differential Revision: D16773657

Pulled By: ppwwyyxx

fbshipit-source-id: 7aba34e150a982998fbad6037156fa6773e68c1c