Adds ExtentTransform, allowing transforms used in CenterNet
Summary:
The ExtentTransform is named after the PIL extent image transform used in its
implementation. It maps a source rectangle in the input image to an output image
of a given size.

This can be used to implement the CenterNet transforms by mapping square source
rectangles onto a 512x512 image, where the source rectangle varies randomly in
its size and center.

Reviewed By: wanyenlo

Differential Revision: D16564535

fbshipit-source-id: ac27d88978c35c55efe4f6231a9320e39d911977