improve logging readability of parameters that will init from scratch
Summary: I recently had a silly bug in which I was training with an R-101 backbone, but loading weights from an R-50 saved model. This would have been easier to catch quickly if the logging of init-from-scratch parameters was clearer. It has already been improved a lot by using the new unfused BN models, which removes a lot of log spam. This diff goes one step further and makes the listing clear (not comma-separated on one line) and colorful to draw attention.

Reviewed By: ppwwyyxx

Differential Revision: D17005450

fbshipit-source-id: ebeb99a126050f972e10faad2cc75731a9ce2141