make `Instances.cat` more general
Summary:
It cannot possibly know how to cat arbitrary types of objects, so add some heuristics.

An alternative is to remove this method.
Currently it is only used once in the codebase (in add_ground_truth_to_proposals).
But there could be other uses (e.g., we used to have it during inference before switching to
`batched_nms`), so I tend to keep it, but wouldn't mind otherwise

Reviewed By: rbgirshick

Differential Revision: D17109619

fbshipit-source-id: 06ce410cd973ba77bc8dd7586c6331b98dc2ce2a