make ANCHOR_GENERATOR use explicit args
Summary:
was planned in https://our.intern.facebook.com/intern/diff/D15741866/?transaction_id=325040368427973

The default anchor generator is pretty small and well-defined.
It's more clear to use explicit arguments.
This also removes the need for a special RetinaNet anchor generator.

Removing the relevant configs is a breaking change and will be done in a later diff.

Reviewed By: rbgirshick

Differential Revision: D17185076

fbshipit-source-id: ae0cc61faf7f092e09ade57f0cd7737acaca944b