Add check for non-empty matches with 0 points
Summary:
Sometimes a situation is encountered where GT boxes are matched with the estimated ones, but none of the GT data falls into the estimated box. Which results in the number of matched boxes `n` being positive and the list of GT points for the estimated box being empty.

So far check has only been valid only for `n == 0`. In case where `n > 0` and the list of GT points is empty, the current code will attempt to concatenate an empty list of Tensors, resulting in an error.

This PR corrects the check to avoid erroneous concatenation.

Reviewed By: ppwwyyxx

Differential Revision: D17257946

fbshipit-source-id: cb011ceca5c5a909fe64b66dcf0937c80d62c306