Allow matching heuristics for non-C2 models
Summary:
Right now, the matching heuristics only applies on C2 models. However, we want it enabled for other
models, because:

Due to the way pytorch names things (that's tied to the architecture), the same parameter (e.g., some layer in the backbone) gets
named differently in different architectures (e.g., FPN vs DC5).

When loading a MOCO-pretrained model, in order to let the same pretrained model file work for both FPN and DC5,
the matching heuristics must be used.

This diff adds a new key in the "pkl" called "matching_heuristic: True". (Feeling hacky).
Any pkl with this field will be name-matched.

Reviewed By: rbgirshick

Differential Revision: D17299049

fbshipit-source-id: 1ce15c488cd11bbba3c5e68a810b562a294272f5