fix bug with DensePose inference on images without detections (#280)
Summary:
Sometimes a situation is encountered where no objects are detected on the image. Which results in  error because of the absence of `pred_densepose` attribute in `instances`  ([line](https://github.com/fairinternal/detectron2/blob/master/projects/DensePose/densepose/evaluator.py#L57)).

Moreover, it is more consistent to add empty DensePose predictions if nothing was detected (in a similar way as it is done for a detection head).

This PR add empty DensePose predictions if nothing was detected. Which prevents from failing when there are no objects detected.
Pull Request resolved: https://github.com/fairinternal/detectron2/pull/280

Reviewed By: rbgirshick

Differential Revision: D17481375

Pulled By: ppwwyyxx

fbshipit-source-id: 8e96181e439eedf00438d4548f6583b76cbd6bd4