Fixed bug where predictions.pred_masks has no attribute numpy (#146)
Summary:
When testing the default `tools/visualise_json_results.py` code an AttributeError was raised on an Instances object. This was in the visualiser.py code where `.numpy()` was not an attribute of `predictions.pred_masks` since it was a list.

This PR wraps the list in the `numpy.asarray` function.
Pull Request resolved: https://github.com/facebookresearch/detectron2/pull/146

Differential Revision: D18085628

Pulled By: ppwwyyxx

fbshipit-source-id: f002270f4b85ea5f9ade3bd77b4fbf5de624cb8d