do not run evaluation in the end of trainer
Summary:
With D19319879, evaluation will run even if the training fails. This is not a reasonable default.
This reverts D19319879.

The issue that D19319879 tried to solve is: when a verification failed, the resumed job would not run evaluation again
and the result verification will be skipped in the resumed job (due to lack of eval metrics).
It causes our integration tests to pass even though the verification has actually failed (as a result we didn't catch a accuracy regression).
This issue can be addressed differently by enforcing the existence of evaluation metrics whenever `cfg.TEST.EXPECTED_RESULTS` is not empty, as done in this diff.

A cleaner solution would require making more objects "checkpointable" so they know where they are after resume.

Reviewed By: rbgirshick

Differential Revision: D19617328

fbshipit-source-id: 105e7648ef406cf7171ae88e9ac7df983a5bc39b