allow per-block arguments in ResNet.make_stage
Summary:
This generalizes `make_stage` to support per-block arguments. It will unify the `first_stride`
argument and also useful for DeepLab's dilation backbone.

Not sure if this would be confusing to use since it might be too magic. Also, we might be over-generalizing it by making "stride" configurable -- that depends on whether a "stage" is defined to be arbitrary list of CNN blocks or a CNN block with the same output spatial size.

Reviewed By: rbgirshick

Differential Revision: D22467495

fbshipit-source-id: fae8eba1d64940aa23ae4eb6780bbcc4ed152ce0