Allow evaluation after resuming training from last iteration
Summary:
Allow evaluation after resuming training from last iteration.
This is useful in the case where the inference/evaluation fails in the training flow, so when the flow retries, it retries the inference/evaluation instead of immediately succeeding.

It slows the work down because we need to relaunch inference separately afterwards.

Reviewed By: ppwwyyxx

Differential Revision: D23783208

fbshipit-source-id: e4404cd984ce8326d7edbf71e994912d2011da16