fix a typo and a bug in ImageList.from_tensors
Summary:
In `ImageList.from_tensors`, shape of each item in `tensors` is (Hi, Wi) or (C_1, ..., C_K, Hi, Wi) where K >= 1 according to the docstring, and from the following code, we can conclude that each item in `tensors` must have the same shape besides width and height (i.e. the last two dimension).

https://github.com/facebookresearch/detectron2/blob/f7b394e059702341c5eb8db2b9a0534779826b4c/detectron2/structures/image_list.py#L117-L121

Accordingly,
https://github.com/facebookresearch/detectron2/blob/f7b394e059702341c5eb8db2b9a0534779826b4c/detectron2/structures/image_list.py#L89-L91

should be
```python
for t in tensors:
    assert isinstance(t, torch.Tensor), type(t)
    assert t.shape[:-2] == tensors[0].shape[:-2], t.shape
```
In addition, there is a typo in docstring in `ImageList.from_tensors`, and this PR fix it.

For your information.

Pull Request resolved: https://github.com/facebookresearch/detectron2/pull/2358

Reviewed By: theschnitz

Differential Revision: D25457842

Pulled By: ppwwyyxx

fbshipit-source-id: 276f5f1a43ccac5f53a44493ad6acde7ed08d88c