don't use a separate stream to gather losses
Summary:
Task in the new stream can start even before the loss is computed.

To do this correctly, it requires creating a cuda event before backward, and wait for the event in the new stream.
The perf difference is tiny, so it's probably not worthwhile so just remove the stream entirely.

Some refs:
https://github.com/pytorch/pytorch/issues/23729

Reviewed By: theschnitz

Differential Revision: D26238335

fbshipit-source-id: 614a5b173861b0c0a2bd1240855f12c19d58b76e