make segm annotation optional
Summary:
Coarse segmentation annotation can be used for training in 2 contexts:
1) together with IUV / CSE annotations within DensePose loss;
2) separately within mask loss.
In the first case, each of the DensePose GT instances must contain coarse segmentation annotation. In the second case, the mask loss is handled separately and not all the instances have to contain coarse segmentation annotations.

To properly support the second case, we make changes to allow training with DensePose loss in the absence of coarse segmentation annotations.

Reviewed By: MarcSzafraniec

Differential Revision: D26327269

fbshipit-source-id: 8d622c63c8083b29bb3efaac3f923ac3fb22a05c