Replace in-place clamp_ with out-place one
Summary:
This PR continues the initiative to replace in-place operators (such as clamp_ , copy_, etc) with out-place version of them. The modification of the graph structure is needed to make it easier to use the model with 3rd party inference frameworks (e.g. TVM, ONNX, etc).

Related discussions:

#2678
ppwwyyxx

Pull Request resolved: https://github.com/facebookresearch/detectron2/pull/2716

Reviewed By: theschnitz

Differential Revision: D26934367

Pulled By: ppwwyyxx

fbshipit-source-id: 12808689556796de3dbb1b483c60b8608f7afa43